proper timeout
On first init, I saw a failure where it actually took 13ms. Okay, let's move the threshold up, how
about nearly 2 orders of magnitude? Like, even if we waited 1 sec at most for a slow to start up
system, that would still be kind of okay, as likely we'd never wait that long.