fix scale imports with similar timestamp
This commit fixes the import from the withings scale where the weight
and the body fat have the same end date, which previously converted
to the same timestamp. Using the timestamp as the primary key in
meins was clearly a mistake and will not repeated in the new flutter
version of the app.